# perf: add channel_id index to objectives_channels