--- a/packages/server-wallet/src/db/migrations/20201121100936_objectives_channels_idx.ts
+++ b/packages/server-wallet/src/db/migrations/20201121100936_objectives_channels_idx.ts
@@ -1,0 +1,11 @@
+import * as Knex from 'knex';
+
+export async function up(knex: Knex): Promise<void> {
+  await knex.schema.alterTable('objectives_channels', function(table) {
+    table.index('channel_id');
+  });
+}
+
+export async function down(_knex: Knex): Promise<void> {
+  //
+}