```
gateway=> EXPLAIN ANALYZE (SELECT * FROM objectives_channels  WHERE channel_id = '0x53bdbb78af3ff8c252aa74c04d5e670ec4d44972e076bd9f4ac4a7a69240e87d');
                                                            QUERY PLAN
-----------------------------------------------------------------------------------------------------------------------------------
 Gather  (cost=1000.00..6266.91 rows=1 width=146) (actual time=0.643..31.103 rows=1 loops=1)
   Workers Planned: 2
   Workers Launched: 2
   ->  Parallel Seq Scan on objectives_channels  (cost=0.00..5266.81 rows=1 width=146) (actual time=12.738..20.625 rows=0 loops=3)
         Filter: ((channel_id)::text = '0x53bdbb78af3ff8c252aa74c04d5e670ec4d44972e076bd9f4ac4a7a69240e87d'::text)
         Rows Removed by Filter: 62470
 Planning Time: 0.052 ms
 Execution Time: 31.120 ms
(8 rows)

gateway=> EXPLAIN ANALYZE (SELECT * FROM objectives_channels  WHERE objective_id = '0x53bdbb78af3ff8c252aa74c04d5e670ec4d44972e076bd9f4ac4a7a69240e87d');
                                                                     QUERY PLAN
-----------------------------------------------------------------------------------------------------------------------------------------------------
 Index Only Scan using objectives_channels_pkey on objectives_channels  (cost=0.55..8.56 rows=1 width=146) (actual time=0.053..0.053 rows=0 loops=1)
   Index Cond: (objective_id = '0x53bdbb78af3ff8c252aa74c04d5e670ec4d44972e076bd9f4ac4a7a69240e87d'::text)
   Heap Fetches: 0
 Planning Time: 0.054 ms
 Execution Time: 0.069 ms
(5 rows)
```